Add files via upload
This PowerShell script provides two main functionalities related to Windows updates: (1) Exporting a list of installed updates to a text file named winupdates list.txt located at C:\temp, and displaying the file's content and location in File Explorer; (2) Checking if a specific Knowledge Base (KB) number is installed on the machine and performing a Bing search for the entered KB number. The script continuously prompts the user to choose between these options, ensuring valid input through a loop, and offers the option to run the program again upon completion.

The first option outputs a list of all installed updates to a file and displays it, while the second option checks the installation status of a specified KB number and searches for it online. After completing a task, the user can choose to run the program again or exit. The script includes input validation to handle invalid choices and prompts the user until a valid choice is made. This ensures a smooth and user-friendly experience.